Test improvements. Updated CHANGELOG.md & README.md. Bumped version '2.3.1-SNAPSHOT' => '2.4.0-SNAPSHOT'
